nyc-taxi-etl-pipeline/pyproject.toml
[project]
name = "nyc-taxi-etl"
version = "0.1.0"
description = "NYC Taxi Trip Data ETL Solution using AWS CDK"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib>=2.185.0",
    "boto3>=1.37.19",
    "constructs>=10.4.2",
    "pytest>=8.3.5",
    "pytest-cov>=4.0.0"
]


[tool.pytest.ini_options]
pythonpath = [
  "."
]